Extract helper for generating local names.
